+++ client/valor/__init__.py
@@ -12,17 +12,14 @@ except importlib_metadata.PackageNotFoundError:
 
 from .client import ClientConnection, connect
 from .coretypes import (
-    Annotation,
     Client,
     Dataset,
-    Datum,
     Evaluation,
     GroundTruth,
-    Label,
     Model,
     Prediction,
 )
-from .schemas import Filter
+from .schemas import Annotation, Datum, Filter, Label
 
 __all__ = [
     "connect",
