+++ client/valor/schemas/symbolic/operators.py
